+++ org.caleydo.view.entourage/src/org/caleydo/view/entourage/GLPathwayGridLayout2.java
@@ -238,10 +238,10 @@ public class GLPathwayGridLayout2 implements IGLLayout {
 		for (PathwayColumn column : copyColumns) {
 			float columnWidth = 0;
 			if (minTotalLevel1Size == 0) {
-				columnWidth = ((float) column.getMinWidth() / (float) totalFixedSize) * getFreeHorizontalSpace(w);
+				columnWidth = (column.getMinWidth() / totalFixedSize) * getFreeHorizontalSpace(w);
 			} else {
 				if (level1Columns.contains(column)) {
-					columnWidth = ((float) column.getMinWidth() / (float) minTotalLevel1Size)
+					columnWidth = (column.getMinWidth() / minTotalLevel1Size)
 							* (getFreeHorizontalSpace(w) - totalFixedSize);
 				} else {
 					columnWidth = column.getMinWidth();
@@ -745,7 +745,7 @@ public class GLPathwayGridLayout2 implements IGLLayout {
 
 		@Override
 		public int compare(PathwayColumn arg0, PathwayColumn arg1) {
-			return arg0.getMinWidth() - arg1.getMinWidth();
+			return (int) (arg0.getMinWidth() - arg1.getMinWidth());
 		}
 
 	}
@@ -754,7 +754,7 @@ public class GLPathwayGridLayout2 implements IGLLayout {
 
 		@Override
 		public int compare(GLPathwayWindow arg0, GLPathwayWindow arg1) {
-			return arg0.getMinHeight() - arg1.getMinHeight();
+			return (int) (arg0.getMinHeight() - arg1.getMinHeight());
 		}
 
 	}
@@ -763,7 +763,7 @@ public class GLPathwayGridLayout2 implements IGLLayout {
 
 		@Override
 		public int compare(GLPathwayWindow arg0, GLPathwayWindow arg1) {
-			return arg0.getMinWidth() - arg1.getMinWidth();
+			return (int) (arg0.getMinWidth() - arg1.getMinWidth());
 		}
 
 	}
@@ -850,10 +850,10 @@ public class GLPathwayGridLayout2 implements IGLLayout {
 			return minHeight;
 		}
 
-		public int getMinWidth() {
-			int maxMinWidth = 0;
+		public float getMinWidth() {
+			float maxMinWidth = 0;
 			for (GLPathwayWindow window : windows) {
-				int minWidth = window.getMinWidth();
+				float minWidth = window.getMinWidth();
 				if (minWidth > maxMinWidth)
 					maxMinWidth = minWidth;
 			}
@@ -890,8 +890,8 @@ public class GLPathwayGridLayout2 implements IGLLayout {
 				float windowHeight = 0;
 
 				if (window.info.getCurrentEmbeddingID() == EEmbeddingID.PATHWAY_LEVEL1) {
-					int minHeight = window.getMinHeight();
-					float factor = (float) minHeight / (float) totalLevel1MinSize;
+					float minHeight = window.getMinHeight();
+					float factor = minHeight / totalLevel1MinSize;
 					windowHeight = factor * (freeSpaceVertical - totalOtherLevelMinSize);
 				} else {
 					windowHeight = window.getMinHeight();
